requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.11.1
Pillow==8.2.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
